requirements.txt
dialogflow
google-auth
asyncinit
aiogcd
yarl
aiohttp
cchardet
aiodns